+++ src/carrier_files/carrier_file_mpeg.cc
@@ -22,7 +22,7 @@ namespace stego_disk
 			auto data_buffer = MemoryBuffer();
 			this->LoadBuffer(data_buffer);
 
-			buffer_.Resize(raw_capacity_);
+			buffer_.Resize(static_cast<std::size_t>(raw_capacity_));
 			buffer_.Clear();
 
 			if (!permutation_->GetSize())
@@ -32,7 +32,7 @@ namespace stego_disk
 
 			for (uint64 i = 0; i < permutation_->GetSize(); ++i)
 			{
-				if (data_buffer[i] & 0x01)
+				if (data_buffer[static_cast<std::size_t>(i)] & 0x01)
 				{
 					SetBitInBufferPermuted(i);
 				}
@@ -49,7 +49,7 @@ namespace stego_disk
 			auto data_buffer = MemoryBuffer();
 			this->LoadBuffer(data_buffer);
 
-			buffer_.Resize(raw_capacity_);
+			buffer_.Resize(static_cast<std::size_t>(raw_capacity_));
 			buffer_.Clear();
 
 			if (!permutation_->GetSize())
@@ -59,7 +59,7 @@ namespace stego_disk
 
 			for (uint64 i = 0; i < permutation_->GetSize(); ++i)
 			{
-				if (data_buffer[i] & 0x01)
+				if (data_buffer[static_cast<std::size_t>(i)] & 0x01)
 				{
 					SetBitInBufferPermuted(i);
 				}
@@ -69,7 +69,7 @@ namespace stego_disk
 
 			for (uint64 i = 0; i < permutation_->GetSize(); ++i)
 			{
-				data_buffer[i] = (data_buffer[i] & 0xFE) | GetBitInBufferPermuted(i);
+				data_buffer[static_cast<std::size_t>(i)] = (data_buffer[static_cast<std::size_t>(i)] & 0xFE) | GetBitInBufferPermuted(i);
 			}
 
 			this->SaveBuffer(data_buffer);
@@ -112,7 +112,7 @@ namespace stego_disk
 
 	void CarrierFileMPEG::LoadBuffer(MemoryBuffer &buffer)
 	{
-		buffer.Resize(raw_capacity_);
+		buffer.Resize(static_cast<std::size_t>(raw_capacity_));
 		buffer.Clear();
 
 		std::size_t byte_position{ 0u }, buff_offset{ 0u };
